+++ admin/paymill_log_entry.php
@@ -1,22 +1,27 @@
 <?php
 
+/**
+ * paymill_log_entry
+ *
+ * @copyright  Copyright (c) 2015 PAYMILL GmbH (https://www.paymill.com)
+ */
 class paymill_log_entry extends paymill_log_abstract
 {
     public function render()
     {
         $this->addTplParam(
-            'debug', 
+            'debug',
             $this->_getEntry(oxConfig::getParameter('entryId'))->paymill_logging__debug->rawValue
         );
-        
+
         return 'paymill_log_entry.tpl';
     }
-    
+
     protected function _getEntry($entryId)
     {
         $data = oxNew('paymill_logging');
         $data->load($entryId);
-        
+
         return $data;
     }
 }