+++ admin/paymill_register_hook.php
@@ -1,26 +1,31 @@
 <?php
 
+/**
+ * paymill_register_hook
+ *
+ * @copyright  Copyright (c) 2015 PAYMILL GmbH (https://www.paymill.com)
+ */
 class paymill_register_hook extends paymill_log_abstract
 {
     public function render()
     {
         parent::render();
-        
+
         $this->addTplParam(
-            'hookUrl', 
+            'hookUrl',
             $this->getConfig()->getSslShopUrl() . 'index.php?cl=paymill_hooks'
         );
-        
+
         return 'paymill_register_hook.tpl';
     }
-    
+
     public function registerHookPoint()
     {
         $webhooks = new Services_Paymill_Webhooks(
             trim(oxConfig::getInstance()->getShopConfVar('PAYMILL_PRIVATEKEY')),
             paymill_util::API_ENDPOINT
         );
-        
+
         $webhooks->create(array(
             "url" => oxConfig::getParameter('hook_url'),
             "event_types" => array('refund.succeeded', 'chargeback.executed')